requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
arrow==1.3.0
bcrypt==4.3.0
binaryornot==0.4.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.3
dotenv==0.9.9
et_xmlfile==2.0.0
flet==0.27.6
flet-cli==0.27.6
flet-desktop==0.27.6
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pypdfium2==4.30.1
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==7.4.2
repath==0.9.0
reportlab==4.4.0
requests==2.32.3
rich==14.0.0
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==4.0.2
XlsxWriter==3.2.3
